requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pandas>=2.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
